Breakdown in modules
- Move to specific dir the implementation
- Breakdown in several modules in order to decouple them
