+++ package.json
@@ -21,10 +21,11 @@
     "body-parser": "^1.19.0",
     "embedme": "^1.22.0",
     "express": "^4.17.1",
+    "json-server": "^0.16.2",
     "local": "^0.3.3",
     "md5": "^2.3.0",
     "nconf": "^0.10.0",
-    "npx": "^10.2.2",
+    "npx": "^3.0.0",
     "sqlite3": "^4.2.0"
   },
   "devDependencies": {
